Moved types into the file where they are used
